+++ Hokkaido.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { Button, View, Text } from 'react-native';
+
+class Hokkaido extends Component {
+    render(){
+        return(
+            <Text>北海道</Text>
+        )
+    }
+}
+
+export default Hokkaido;
